fix: silence ~80+ test warnings from own code
- Use jnp.float32 instead of jnp.float64 for HBV triangular routing
  weights (model.py, routing.py) since jax_enable_x64 is not enabled
- Strip tzinfo before numpy datetime64 conversion in pi_xml.py
  (datetimes already normalized to UTC)
- Replace parse_dates=True with explicit pd.to_datetime(format='mixed')
  in base evaluator and observation loader to avoid ambiguous parsing
- Guard matplotlib legend call with warnings.catch_warnings() in
  ignition.py to suppress PatchCollection handle warning

+++ src/symfluence/models/hbv/routing.py
@@ -389,8 +389,8 @@ def route_network_step_jax(
         node = topo_order[node_idx]
 
         # Sum upstream contributions
-        upstream_sum = jnp.float64(0.0)
-        prev_upstream_sum = jnp.float64(0.0)
+        upstream_sum = jnp.float32(0.0)
+        prev_upstream_sum = jnp.float32(0.0)
 
         # Add contributions from all upstream nodes
         def add_upstream(i, sums):
